\name{EmpiricalIO-package}
\alias{EmpiricalIO-package}
\alias{EmpiricalIO}
\docType{package}
\title{
\packageTitle{EmpiricalIO}
}
\description{
\packageDescription{EmpiricalIO}
}
\details{
%%  ~~ An overview of how to use the package, ~~
%%  ~~ including the most important functions ~~
}
\author{
\packageAuthor{EmpiricalIO}

Maintainer: \packageMaintainer{EmpiricalIO}
}
\references{
%%  ~~ Literature or other references for background information ~~
}
\keyword{package}

%% Uncomment below to imitate parts of library(help = EmpiricalIO)
%\section{The \file{DESCRIPTION} File}{\packageDESCRIPTION{EmpiricalIO}}
%\section{Documentation Index}{\packageIndices{EmpiricalIO}}

\seealso{
%%  ~~ Optional links to other man pages, e.g. ~~
%%  ~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
%%  ~~ Optional simple examples of the most important functions ~~
}
