\name{EmpiricalIO-package}
\alias{EmpiricalIO-package}
\alias{EmpiricalIO}
\docType{package}
\title{
\packageTitle{EmpiricalIO}
}
\description{
\packageDescription{EmpiricalIO}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EmpiricalIO}
\packageIndices{EmpiricalIO}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{EmpiricalIO}

Maintainer: \packageMaintainer{EmpiricalIO}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from ~~
~~ file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
